resource.go, util.go Refactor Create method on resource
Tests were failing due to pointer issue, resulting from failing to create a resource.

First fix was to change mongo functions in util.go to exported function, and then change their use in all of the models.

Suspected issue was in ErrorDocumentExists implementation, appears not to be the case.

Turns out it was the resource Create method was calling MongoCreateOne, creating a pointer issue, I suspect from the argument being an interface{} instead of an interface supporting MarshalBSON